==> notebooks/6_preprocessing.ipynb <==
# Ideas
Search for words that appear most commonly together with others (e.g., Trump, Biden, president, etc.)

from nltk.corpus import stopwords
from tqdm import tqdm

import numpy as np
import pandas as pd
import warnings
import spacy

warnings.filterwarnings("ignore")
en_stopwords = stopwords.words('english')
nlp = spacy.load('en_core_web_lg')

fillers = [
    'like',
    'um',
    'uh',
    'uhm',
    'you know',
    'well',
    'so',
    'basically',
    'actually',
    'literally',
    'kind of',
    'sort of',
    'kinda',
    'sorta',
    'really',
    'totally',
    'seriously',
    'just',
    'maybe',
    'probably'
]

data = pd.read_csv('../data/processed/cleaned_transcripted_dataset.csv')
print(data.shape)
data.head()

# Preprocessing

data['full_text'] = data['video_desc'].fillna('') + ' ' + data['transcription'] + ' ' + data['stickers_on_video'].fillna('')
data['full_text'] = data['full_text'].str.lower()

print(data.full_text.iloc[1])

# Lemmatization

lemmas = []
for doc in tqdm(nlp.pipe(data['full_text'], disable=["tok2vec", "tagger", "parser", "attribute_ruler"], n_process=-1), total=len(data['full_text'])):
    doc_lemmas = []
    for token in doc:
        str_token = str(token)
        if not token.is_punct and str_token not in en_stopwords:
            doc_lemmas.append(token.lemma_)
    lemmas.append(' '.join(doc_lemmas))

data['full_text_lemmas'] = lemmas

data.full_text_lemmas.head(5)

# Removing fillers

for filler in fillers:
    filler = ' {} '.format(filler)
    data['full_text_lemmas'] = data['full_text_lemmas'].str.replace(filler, ' ')

# Saving dataset

data.to_csv('../data/processed/preprocessed_lemma_dataset.csv', index=False)
